Refactored CampusLocations to BuildingInfo
Removing duplicate class for the same feature
