--- app/src/main/java/com/example/scoutconcordia/CampusLocations.java
@@ -1,1110 +1,0 @@
-package com.example.scoutconcordia;
-
-public class CampusLocations
-{
-    /**
-     *  This method is used to obtain all the locations on the SGW campus
-     */
-    public static double[][][] getSGWLocations()
-    {
-        // (B) 2160, Bishop
-        double[][] lBbuilding =
-        {
-            {45.497741, -73.579635},
-            {45.497921, -73.579455},
-            {45.497884, -73.579376},
-            {45.497708, -73.579562}
-        };
-    
-        // (CI) 2149, Mackay
-        double[][] lCIbuilding =
-        {
-            {45.497406, -73.580015},
-            {45.497583, -73.579837},
-            {45.497546, -73.579760},
-            {45.497365, -73.579932}
-        };
-    
-        // (CL) 1665, Sainte-Catherine O.
-        double[][] lCLbuilding =
-        {
-            {45.494165, -73.579656},
-            {45.494471, -73.579282},
-            {45.494260, -73.578937},
-            {45.494035, -73.579214},
-            {45.494012, -73.579249},
-            {45.493989, -73.579293},
-            {45.493978, -73.579332},
-            {45.493978, -73.579332}
-        };
-    
-        // (D) 2140, Bishop
-        double[][] lDbuilding =
-        {
-            {45.497742, -73.579411},
-            {45.497847, -73.579308},
-            {45.497808, -73.579222},
-            {45.497707, -73.579342}
-        };
-    
-    
-        // (EN) 2070, Mackay
-        double[][] lENbuilding =
-        {
-            {45.496699, -73.579784},
-            {45.496931, -73.579553},
-            {45.496891, -73.579469},
-            {45.496788, -73.579574},
-            {45.496802, -73.579606},
-            {45.496676, -73.579733},
-        };
-    
-        // (ER) 2155, Guy
-        double[][] lERbuilding =
-        {
-            {45.496260, -73.580351},
-            {45.496677, -73.579968},
-            {45.496522, -73.579631},
-            {45.496216, -73.579905},
-            {45.496247, -73.579972},
-            {45.496162, -73.580068},
-        };
-    
-        // (EV) Engineering, Computer Science and Visual Arts Integrated Complex. 1515, Sainte-Catherine O.
-        double[][] lEVbuilding =
-        {
-            {45.495595, -73.578762},
-            {45.495863, -73.578490},
-            {45.495668, -73.578073},
-            {45.496047, -73.577708},
-            {45.495831, -73.577249},
-            {45.495539, -73.577530},
-            {45.495550, -73.577555},
-            {45.495462, -73.577639},
-            {45.495447, -73.577609},
-            {45.495355, -73.577695},
-            {45.495353, -73.577742},
-            {45.495237, -73.577877},
-            {45.495234, -73.577913},
-            {45.495248, -73.577914},
-            {45.495264, -73.578001},
-            {45.495247, -73.578021}
-        };
-    
-    
-        // (FA) 2060, Mackay
-        double[][] lFAbuilding =
-        {
-            {45.496777, -73.579586},
-            {45.496871, -73.579488},
-            {45.496835, -73.579402},
-            {45.496735, -73.579503}
-        };
-    
-        // (FB) Faubourg Building 1250, Guy. 1600 Sainte-Catherine O.
-        double[][] lFBbuilding =
-        {
-            {45.494696, -73.578039},
-            {45.494912, -73.577785},
-            {45.494869, -73.577713},
-            {45.494876, -73.577704},
-            {45.494835, -73.577633},
-            {45.494842, -73.577624},
-            {45.494798, -73.577549},
-            {45.494806, -73.577541},
-            {45.494763, -73.577465},
-            {45.494774, -73.577453},
-            {45.494692, -73.577310},
-            {45.494700, -73.577299},
-            {45.494654, -73.577217},
-            {45.494396, -73.577521},
-        };
-    
-    
-        // (FG) Faubourg Sainte-Catherine Building 1610, Sainte-Catherine O.
-        double[][] lFGbuilding =
-        {
-            {45.493822, -73.579067},
-            {45.494294, -73.578512},
-            {45.494300, -73.578523},
-            {45.494369, -73.578444},
-            {45.494361, -73.578426},
-            {45.494694, -73.578039},
-            {45.494452, -73.577617},
-            {45.494387, -73.577692},
-            {45.494426, -73.577762},
-            {45.494392, -73.577805},
-            {45.494371, -73.577768},
-            {45.494186, -73.577987},
-            {45.494202, -73.578018},
-            {45.494111, -73.578126},
-            {45.494103, -73.578113},
-            {45.493911, -73.578342},
-            {45.493922, -73.578364},
-            {45.493891, -73.578400},
-            {45.493882, -73.578385},
-            {45.493833, -73.578440},
-            {45.493847, -73.578464},
-            {45.493624, -73.578727},
-        };
-    
-        // (GA) Grey Nuns Annex 1211-1215, St-Mathieu
-        double[][] lGAbuilding =
-        {
-            {45.493850, -73.578355},
-            {45.494134, -73.578011},
-            {45.494119, -73.577984},
-            {45.494345, -73.577736},
-            {45.494284, -73.577622},
-            {45.494054, -73.577870},
-            {45.494075, -73.577908},
-            {45.493790, -73.578254}
-        };
-    
-        // (GM) Guy-De Maisonneuve Building
-        double[][] lGMbuilding =
-        {
-            {45.495781, -73.579147},
-            {45.496128, -73.578806},
-            {45.496111, -73.578767},
-            {45.496109, -73.578767},
-            {45.495947, -73.578435},
-            {45.495616, -73.578744},
-            {45.495779, -73.579090},
-            {45.495763, -73.579108}
-        };
-    
-        // (GN) Grey Nuns Building 1190, Guy. 1175, St-Mathieu
-        double[][] lGNbuilding =
-        {
-            {45.492593, -73.576534},
-            {45.492614, -73.576577},
-            {45.492607, -73.576584},
-            {45.492746, -73.576869},
-            {45.492727, -73.576873},
-            {45.492676, -73.576923},
-            {45.492677, -73.576965},
-            {45.492710, -73.577033},
-            {45.492773, -73.576973},
-            {45.492794, -73.576977},
-            {45.492810, -73.576962},
-            {45.492896, -73.577141},
-            {45.492838, -73.577198},
-            {45.492925, -73.577379},
-            {45.492993, -73.577314},
-            {45.493088, -73.577511},
-            {45.493198, -73.577401},
-            {45.493108, -73.577213},
-            {45.493363, -73.576961},
-            {45.493353, -73.576936},
-            {45.493438, -73.576853},
-            {45.493478, -73.576935},
-            {45.493450, -73.576963},
-            {45.493529, -73.577128},
-            {45.493576, -73.577081},
-            {45.493601, -73.577131},
-            {45.493579, -73.577153},
-            {45.493634, -73.577267},
-            {45.493746, -73.577157},
-            {45.493671, -73.576999},
-            {45.493727, -73.576944},
-            {45.493667, -73.576822},
-            {45.493613, -73.576873},
-            {45.493551, -73.576747},
-            {45.493838, -73.576465},
-            {45.493837, -73.576669},
-            {45.493717, -73.576784},
-            {45.493902, -73.577173},
-            {45.494065, -73.577015},
-            {45.493897, -73.576660},
-            {45.493929, -73.576472},
-            {45.494193, -73.577020},
-            {45.493868, -73.577339},
-            {45.493975, -73.577559},
-            {45.494124, -73.577413},
-            {45.494094, -73.577348},
-            {45.494391, -73.577055},
-            {45.494018, -73.576282},
-            {45.494131, -73.576168},
-            {45.494043, -73.575991},
-            {45.493933, -73.576093},
-            {45.493713, -73.575638},
-            {45.493570, -73.575782},
-            {45.493792, -73.576245},
-            {45.493492, -73.576540},
-            {45.493472, -73.576497},
-            {45.493340, -73.576624},
-            {45.493364, -73.576673},
-            {45.493024, -73.577006},
-            {45.492941, -73.576833},
-            {45.492948, -73.576794},
-            {45.492926, -73.576747},
-            {45.492899, -73.576743},
-            {45.492733, -73.576396}
-        };
-    
-        // (GS) 1538, Sherbrooke
-        double[][] lGSbuilding =
-        {
-            {45.496627, -73.581460},
-            {45.496785, -73.581300},
-            {45.496711, -73.581126},
-            {45.496653, -73.581171},
-            {45.496518, -73.580803},
-            {45.496477, -73.580844},
-            {45.496487, -73.580872},
-            {45.496415, -73.580941},
-            {45.496578, -73.581436}
-        };
-    
-        // (H) Henry F. Hall Building. 1455, De Maisonneuve O.
-        double[][] lHbuilding =
-        {
-            {45.497163, -73.579557},
-            {45.497718, -73.579039},
-            {45.497374, -73.578319},
-            {45.496820, -73.578841}
-        };
-    
-        // (K) 2150, Bishop
-        double[][] lKbuilding =
-        {
-            {45.497627, -73.579634},
-            {45.497881, -73.579379},
-            {45.497847, -73.579307},
-            {45.497754, -73.579401},
-            {45.497772, -73.579438},
-            {45.497697, -73.579510},
-            {45.497685, -73.579487},
-            {45.497597, -73.579571}
-        };
-    
-        // (LB) J.W McConnell Building. 1400, De Maisonneuve O.
-        double[][] lLBbuilding =
-        {
-            {45.496706, -73.578530},
-            {45.496729, -73.578578},
-            {45.496887, -73.578417},
-            {45.496870, -73.578373},
-            {45.496913, -73.578331},
-            {45.496895, -73.578293},
-            {45.496940, -73.578247},
-            {45.496964, -73.578296},
-            {45.497001, -73.578258},
-            {45.497018, -73.578297},
-            {45.497270, -73.578049},
-            {45.497143, -73.577817},
-            {45.497109, -73.577849},
-            {45.497085, -73.577800},
-            {45.497078, -73.577808},
-            {45.497070, -73.577792},
-            {45.497121, -73.577744},
-            {45.497040, -73.577580},
-            {45.496994, -73.577632},
-            {45.496988, -73.577621},
-            {45.496999, -73.577610},
-            {45.496978, -73.577567},
-            {45.497013, -73.577535},
-            {45.496892, -73.577293},
-            {45.496616, -73.577556},
-            {45.496637, -73.577602},
-            {45.496583, -73.577651},
-            {45.496496, -73.577465},
-            {45.496256, -73.577698},
-            {45.496668, -73.578568}
-        };
-    
-        // (LD) 1424, Bishop
-        double[][] lLDbuilding =
-        {
-            {45.496564, -73.577470},
-            {45.496858, -73.577170},
-            {45.496818, -73.577086},
-            {45.496522, -73.577386}
-        };
-    
-        // (LS) 1535, De Maisonneuve O.
-        double[][]  lLSbuilding =
-        {
-            {45.496200, -73.579633},
-            {45.496283, -73.579551},
-            {45.496339, -73.579731},
-            {45.496496, -73.579564},
-            {45.496327, -73.579162},
-            {45.496095, -73.579363}
-        };
-    
-        // (M) 2135, Mackay
-        double[][] lMbuilding =
-        {
-            {45.497323, -73.579844},
-            {45.497426, -73.579734},
-            {45.497390, -73.579683},
-            {45.497290, -73.579780}
-        };
-    
-        // (MB) John Molson Building. 1450, Guy
-        double[][] lMBbuilding =
-        {
-            {45.495358, -73.579367},
-            {45.495519, -73.579200},
-            {45.495439, -73.578960},
-            {45.495187, -73.578525},
-            {45.495004, -73.578736},
-            {45.495037, -73.578790},
-            {45.495006, -73.578824},
-            {45.495166, -73.579171},
-            {45.495221, -73.579115}
-        };
-    
-        // (MI) 2130, Bishop
-        double[][] lMIbuilding =
-        {
-            {45.497647, -73.579403},
-            {45.497808, -73.579222},
-            {45.497778, -73.579167},
-            {45.497683, -73.579260},
-            {45.497694, -73.579284},
-            {45.497623, -73.579354}
-        };
-    
-        // (MU) 2170, Bishop
-        double[][] lMUbuilding =
-        {
-            {45.497785, -73.579703},
-            {45.497964, -73.579533},
-            {45.497921, -73.579455},
-            {45.497749, -73.579628}
-        };
-    
-        // (P) 2020, Mackay
-        double[][] lPbuilding =
-        {
-            {45.496625, -73.579262},
-            {45.496723, -73.579170},
-            {45.496677, -73.579083},
-            {45.496578, -73.579171}
-        };
-    
-        // (PR) 2100, Mackay
-        double[][] lPRbuilding =
-        {
-            {45.496946, -73.579938},
-            {45.497029, -73.579852},
-            {45.496980, -73.579748},
-            {45.496895, -73.579833}
-        };
-    
-        // (Q) 2010, Mackay
-        double[][] lQbuilding =
-        {
-            {45.496578, -73.579170},
-            {45.496675, -73.579082},
-            {45.496644, -73.579009},
-            {45.496553, -73.579120}
-        };
-    
-        // (R) 2050, Mackay
-        double[][] lRbuilding =
-        {
-            {45.496745, -73.579492},
-            {45.496837, -73.579400},
-            {45.496802, -73.579329},
-            {45.496711, -73.579420}
-        };
-    
-        // (RR) 2040, Mackay
-        double[][] lRRbuilding =
-        {
-            {45.496651, -73.579480},
-            {45.496796, -73.579331},
-            {45.496759, -73.579254},
-            {45.496612, -73.579401}
-        };
-    
-        // (S) 2145, Mackay
-        double[][] lSbuilding =
-        {
-            {45.497363, -73.579927},
-            {45.497491, -73.579799},
-            {45.497460, -73.579753},
-            {45.497437, -73.579773},
-            {45.497426, -73.579734},
-            {45.497323, -73.579844}
-        };
-    
-        // (SB) Samuel Bronfman Building. 1590, Docteur-Penfield
-        double[][] lSBbuilding =
-        {
-            {45.496435, -73.586209},
-            {45.496501, -73.586208},
-            {45.496506, -73.586135},
-            {45.496521, -73.586136},
-            {45.496519, -73.586168},
-            {45.496538, -73.586212},
-            {45.496683, -73.586087},
-            {45.496652, -73.586014},
-            {45.496658, -73.586008},
-            {45.496582, -73.585830},
-            {45.496572, -73.585835},
-            {45.496553, -73.585792},
-            {45.496509, -73.585792},
-            {45.496511, -73.585779},
-            {45.496491, -73.585739},
-            {45.496467, -73.585737}
-        };
-    
-        // (T) 2030, Mackay
-        double[][] lTbuilding =
-        {
-            {45.496668, -73.579345},
-            {45.496751, -73.579260},
-            {45.496711, -73.579183},
-            {45.496627, -73.579264}
-        };
-    
-        // (TD) Toronto-Dominion Building
-        double[][] lTDbuilding =
-        {
-            {45.495128, -73.578501},
-            {45.495189, -73.578427},
-            {45.495037, -73.578072},
-            {45.494943, -73.578177},
-            {45.495023, -73.578325},
-            {45.495043, -73.578300},
-            {45.495065, -73.578341},
-            {45.495047, -73.578364}
-        };
-    
-        // (V) 2110, Mackay
-        double[][] lVbuilding =
-        {
-            {45.496984, -73.580018},
-            {45.497089, -73.579916},
-            {45.497048, -73.579834},
-            {45.496946, -73.579938}
-        };
-    
-        // (VA) Visual Arts Building. 1395, Rene-Levesque O.
-        double[][] lVAbuilding =
-        {
-            {45.495401, -73.573767},
-            {45.495669, -73.574309},
-            {45.496186, -73.573795},
-            {45.496069, -73.573560},
-            {45.495817, -73.573810},
-            {45.495675, -73.573490}
-        };
-    
-        // (X) 2080, Mackay
-        double[][] lXbuilding =
-        {
-            {45.496857, -73.579758},
-            {45.496949, -73.579665},
-            {45.496907, -73.579578},
-            {45.496815, -73.579671}
-        };
-    
-        // (Z) 2090, Mackay
-        double[][] lZbuilding =
-        {
-            {45.496911, -73.579816},
-            {45.496986, -73.579741},
-            {45.496949, -73.579665},
-            {45.496850, -73.579763},
-            {45.496874, -73.579812},
-            {45.496897, -73.579791}
-        };
-    
-        return new double[][][]{
-        lBbuilding, lCIbuilding, lCLbuilding, lDbuilding, lENbuilding, lERbuilding,
-        lEVbuilding, lFAbuilding, lFBbuilding, lFGbuilding, lGAbuilding, lGMbuilding,
-        lGNbuilding, lGSbuilding, lHbuilding, lKbuilding, lLBbuilding, lLDbuilding,
-        lLSbuilding, lMbuilding, lMBbuilding, lMIbuilding, lMUbuilding, lPbuilding,
-        lPRbuilding, lQbuilding, lRbuilding, lRRbuilding, lSbuilding, lSBbuilding,
-        lTbuilding, lTDbuilding, lVbuilding, lVAbuilding, lXbuilding, lZbuilding};
-    }
-    
-    /**
-     *  This method is used to obtain all the locations on the SGW campus
-     */
-    public static double[][][] getLoyolaLocations()
-    {
-        // (AD) 7141, Sherbrooke O. (Administration Building)
-        double[][] lADbuilding =
-        {
-            {45.457911, -73.640124},
-            {45.457985, -73.640067},
-            {45.457962, -73.640006},
-            {45.458086, -73.639912},
-            {45.458089, -73.639920},
-            {45.458154, -73.639869},
-            {45.458151, -73.639862},
-            {45.458277, -73.639767},
-            {45.458298, -73.639821},
-            {45.458371, -73.639765},
-            {45.458327, -73.639649},
-            {45.458331, -73.639644},
-            {45.458313, -73.639595},
-            {45.458307, -73.639598},
-            {45.458259, -73.639469},
-            {45.458242, -73.639483},
-            {45.458230, -73.639476},
-            {45.458210, -73.639494},
-            {45.458207, -73.639510},
-            {45.458183, -73.639527},
-            {45.458200, -73.639574},
-            {45.458191, -73.639582},
-            {45.458203, -73.639617},
-            {45.458097, -73.639697},
-            {45.458086, -73.639691},
-            {45.458065, -73.639634},
-            {45.458060, -73.639638},
-            {45.458040, -73.639581},
-            {45.457983, -73.639626},
-            {45.458004, -73.639679},
-            {45.458000, -73.639682},
-            {45.458022, -73.639743},
-            {45.458023, -73.639755},
-            {45.457915, -73.639837},
-            {45.457903, -73.639808},
-            {45.457893, -73.639817},
-            {45.457875, -73.639771},
-            {45.457846, -73.639791},
-            {45.457841, -73.639788},
-            {45.457822, -73.639795},
-            {45.457819, -73.639812},
-            {45.457799, -73.639827}
-        };
-    
-        // (BB) 3502, Belmore
-        double[][] lBBbuilding =
-        {
-            {45.459731, -73.639267},
-            {45.459828, -73.639192},
-            {45.459797, -73.639118},
-            {45.459702, -73.639192}
-        };
-    
-        // (BH) 3500, Belmore
-        double[][] lBHbuilding =
-        {
-            {45.459702, -73.639192},
-            {45.459797, -73.639118},
-            {45.459770, -73.639044},
-            {45.459676, -73.639116}
-        };
-    
-        // (CC) 7141, Sherbrooke O. (Central Building)
-        double[][] lCCbuilding =
-        {
-            {45.458379, -73.640794},
-            {45.458518, -73.640684},
-            {45.458429, -73.640450},
-            {45.458436, -73.640444},
-            {45.458321, -73.640142},
-            {45.458313, -73.640146},
-            {45.458221, -73.639906},
-            {45.458166, -73.639948},
-            {45.458140, -73.639881},
-            {45.458108, -73.639906},
-            {45.458122, -73.639944},
-            {45.458095, -73.639960},
-            {45.458104, -73.639995},
-            {45.458081, -73.640015},
-            {45.458173, -73.640255},
-            {45.458167, -73.640262},
-            {45.458280, -73.640563},
-            {45.458289, -73.640557}
-        };
-    
-        // (CJ) 7141, Sherbrooke O. (Communication Studies and Journalism Building)
-        double[][] lCJbuilding =
-        {
-            {45.457333, -73.640717},
-            {45.457597, -73.640503},
-            {45.457649, -73.640631},
-            {45.457831, -73.640486},
-            {45.457755, -73.640291},
-            {45.457726, -73.640315},
-            {45.457622, -73.640045},
-            {45.457484, -73.640149},
-            {45.457435, -73.640030},
-            {45.457446, -73.639945},
-            {45.457464, -73.639954},
-            {45.457479, -73.639819},
-            {45.457431, -73.639772},
-            {45.457387, -73.639760},
-            {45.457359, -73.639761},
-            {45.457333, -73.639769},
-            {45.457305, -73.639784},
-            {45.457279, -73.639802},
-            {45.457259, -73.639828},
-            {45.457229, -73.639883},
-            {45.457211, -73.639984},
-            {45.457213, -73.640018},
-            {45.457305, -73.640073},
-            {45.457311, -73.640049},
-            {45.457361, -73.640076},
-            {45.457411, -73.640206},
-            {45.457175, -73.640393},
-            {45.457280, -73.640658},
-            {45.457304, -73.640639}
-        };
-    
-        // (DO) 7200, Sherbrooke O. (Stinger Dome)
-        double[][] lDObuilding =
-        {
-            {45.457389, -73.637103},
-            {45.458364, -73.635969},
-            {45.457932, -73.635218},
-            {45.456959, -73.636361}
-        };
-    
-        // (FC) 7141, Sherbrooke O. ( F.C. Smith Building)
-        double[][] lFCbuilding =
-        {
-            {45.458599, -73.639588},
-            {45.458627, -73.639665},
-            {45.458665, -73.639683},
-            {45.458741, -73.639624},
-            {45.458750, -73.639571},
-            {45.458721, -73.639493},
-            {45.458734, -73.639485},
-            {45.458709, -73.639417},
-            {45.458717, -73.639412},
-            {45.458702, -73.639372},
-            {45.458720, -73.639360},
-            {45.458717, -73.639352},
-            {45.458743, -73.639333},
-            {45.458727, -73.639295},
-            {45.458702, -73.639315},
-            {45.458697, -73.639303},
-            {45.458680, -73.639316},
-            {45.458670, -73.639290},
-            {45.458661, -73.639295},
-            {45.458601, -73.639136},
-            {45.458610, -73.639128},
-            {45.458615, -73.639095},
-            {45.458603, -73.639066},
-            {45.458582, -73.639058},
-            {45.458575, -73.639065},
-            {45.458569, -73.639051},
-            {45.458558, -73.639058},
-            {45.458537, -73.639003},
-            {45.458520, -73.639016},
-            {45.458507, -73.638982},
-            {45.458528, -73.638965},
-            {45.458518, -73.638935},
-            {45.458382, -73.639040},
-            {45.458392, -73.639070},
-            {45.458415, -73.639052},
-            {45.458427, -73.639086},
-            {45.458410, -73.639102},
-            {45.458430, -73.639154},
-            {45.458424, -73.639161},
-            {45.458432, -73.639181},
-            {45.458419, -73.639192},
-            {45.458442, -73.639255},
-            {45.458457, -73.639244},
-            {45.458517, -73.639408},
-            {45.458511, -73.639411},
-            {45.458514, -73.639423},
-            {45.458487, -73.639442},
-            {45.458512, -73.639520},
-            {45.458525, -73.639509},
-            {45.458535, -73.639540},
-            {45.458531, -73.639561},
-            {45.458539, -73.639585},
-            {45.458556, -73.639594},
-            {45.458580, -73.639574},
-            {45.458590, -73.639597}
-        };
-    
-        // (GE) 7141, Sherbrooke O. (Centre for Structural and Functional Genomics)
-        double[][] lGEbuilding =
-        {
-            {45.456945, -73.640742},
-            {45.457174, -73.640570},
-            {45.457130, -73.640452},
-            {45.457142, -73.640443},
-            {45.457040, -73.640164},
-            {45.456798, -73.640346},
-            {45.456896, -73.640612},
-            {45.456872, -73.640630},
-            {45.456894, -73.640689},
-            {45.456918, -73.640671}
-        };
-    
-        // (HA) 7141, Sherbrooke O. (Hingston Hall, wing HA)
-        double[][] lHAbuilding =
-        {
-            {45.459410, -73.641485},
-            {45.459432, -73.641543},
-            {45.459676, -73.641352},
-            {45.459652, -73.641290},
-            {45.459663, -73.641281},
-            {45.459529, -73.640931},
-            {45.459515, -73.640941},
-            {45.459492, -73.640882},
-            {45.459252, -73.641070},
-            {45.459276, -73.641133},
-            {45.459260, -73.641146},
-            {45.459395, -73.641495}
-        };
-    
-        // (HB) 7141, Sherbrooke O. (Hingston Hall, wing HB)
-        double[][] lHBbuilding =
-        {
-            {45.459140, -73.642244},
-            {45.459161, -73.642303},
-            {45.459360, -73.642145},
-            {45.459339, -73.642089},
-            {45.459362, -73.642070},
-            {45.459358, -73.642057},
-            {45.459502, -73.641943},
-            {45.459527, -73.642003},
-            {45.459556, -73.641978},
-            {45.459368, -73.641496},
-            {45.458972, -73.641806},
-            {45.458997, -73.641878},
-            {45.458963, -73.641906},
-            {45.459105, -73.642271}
-        };
-    
-        // (HC) 7141, Sherbrooke O. (Hingston Hall, wing HC)
-        double[][] lHCbuilding =
-        {
-            {45.459623, -73.642271},
-            {45.459794, -73.642134},
-            {45.459781, -73.642100},
-            {45.459893, -73.642010},
-            {45.459825, -73.641833},
-            {45.459713, -73.641920},
-            {45.459698, -73.641885},
-            {45.459525, -73.642023}
-        };
-    
-        // (HU) 7141, Sherbrooke O. (Applied Science hub)
-        double[][] lHUbuilding =
-        {
-            {45.458290, -73.641709},
-            {45.458530, -73.642346},
-            {45.458847, -73.642096},
-            {45.458625, -73.641470}
-        };
-    
-        // (JR) 7141, Sherbrooke O. (Jesuit Residence)
-        double[][] lJRbuilding =
-        {
-            {45.458495, -73.643402},
-            {45.458559, -73.643348},
-            {45.458566, -73.643364},
-            {45.458634, -73.643308},
-            {45.458606, -73.643236},
-            {45.458624, -73.643221},
-            {45.458583, -73.643119},
-            {45.458570, -73.643130},
-            {45.458539, -73.643055},
-            {45.458484, -73.643099},
-            {45.458489, -73.643112},
-            {45.458478, -73.643121},
-            {45.458469, -73.643096},
-            {45.458397, -73.643153},
-            {45.458427, -73.643227},
-            {45.458410, -73.643242},
-            {45.458448, -73.643343},
-            {45.458466, -73.643329}
-        };
-    
-        // (PC) 7200, Sherbrooke O. (PERFORM Centre)
-        double[][] lPCbuilding =
-        {
-            {45.457013, -73.637885},
-            {45.457279, -73.637674},
-            {45.456942, -73.636805},
-            {45.456675, -73.637024}
-        };
-    
-        // (PS) 7141, Sherbrooke O. (Physical Services Building)
-        double[][] lPSbuilding =
-        {
-            {45.459706, -73.640312},
-            {45.459851, -73.640198},
-            {45.459864, -73.640228},
-            {45.459981, -73.640134},
-            {45.459664, -73.639317},
-            {45.459624, -73.639349},
-            {45.459576, -73.639229},
-            {45.459287, -73.639457},
-            {45.459333, -73.639579},
-            {45.459403, -73.639525},
-            {45.459444, -73.639631},
-            {45.459414, -73.639655},
-            {45.459607, -73.640158},
-            {45.459638, -73.640135}
-        };
-    
-        // (PT) 7141, Sherbrooke O. (Oscar Peterson Concert Hall)
-        double[][] lPTbuilding =
-        {
-            {45.459372, -73.639253},
-            {45.459505, -73.639149},
-            {45.459470, -73.639054},
-            {45.459458, -73.639071},
-            {45.459312, -73.638696},
-            {45.459170, -73.638811},
-            {45.459314, -73.639189},
-            {45.459322, -73.639183},
-            {45.459340, -73.639234},
-            {45.459361, -73.639220}
-        };
-    
-        // (PY) 7141, Sherbrooke O. (Psychology Building)
-        double[][] lPYbuilding =
-        {
-            {45.458848, -73.640835},
-            {45.459180, -73.640573},
-            {45.459200, -73.640623},
-            {45.459283, -73.640560},
-            {45.459233, -73.640432},
-            {45.459238, -73.640428},
-            {45.459216, -73.640372},
-            {45.459212, -73.640375},
-            {45.459120, -73.640134},
-            {45.459087, -73.640159},
-            {45.459084, -73.640152},
-            {45.459053, -73.640179},
-            {45.459057, -73.640183},
-            {45.459024, -73.640209},
-            {45.459029, -73.640222},
-            {45.458996, -73.640248},
-            {45.458985, -73.640220},
-            {45.458758, -73.640398},
-            {45.458764, -73.640418},
-            {45.458731, -73.640446},
-            {45.458823, -73.640695},
-            {45.458801, -73.640713}
-        };
-    
-        // (RA) 7200, Sherbrooke O. (Recreation and Athletics Complex)
-        double[][] lRAbuilding =
-        {
-            {45.456887, -73.638561},
-            {45.457018, -73.638455},
-            {45.457031, -73.638486},
-            {45.457055, -73.638466},
-            {45.457042, -73.638432},
-            {45.457158, -73.638339},
-            {45.457036, -73.638033},
-            {45.457008, -73.638055},
-            {45.456957, -73.637927},
-            {45.457027, -73.637872},
-            {45.456726, -73.637102},
-            {45.456682, -73.636989},
-            {45.456415, -73.637368},
-            {45.456397, -73.637385},
-            {45.456692, -73.638140},
-            {45.456794, -73.638060},
-            {45.456844, -73.638188},
-            {45.456774, -73.638244},
-            {45.456802, -73.638316},
-            {45.456767, -73.638345},
-            {45.456782, -73.638386},
-            {45.456794, -73.638378},
-            {45.456797, -73.638387},
-            {45.456786, -73.638398},
-            {45.456803, -73.638441},
-            {45.456833, -73.638417}
-        };
-    
-        // (RF) 7141, Sherbrooke O. (Loyola Jesuit Hall and Conference Centre)
-        double[][] lRFbuilding =
-        {
-            {45.458508, -73.641375},
-            {45.458641, -73.641270},
-            {45.458647, -73.641284},
-            {45.458806, -73.641160},
-            {45.458801, -73.641145},
-            {45.458821, -73.641129},
-            {45.458785, -73.641036},
-            {45.458763, -73.641050},
-            {45.458692, -73.640854},
-            {45.458699, -73.640851},
-            {45.458683, -73.640807},
-            {45.458587, -73.640880},
-            {45.458539, -73.640756},
-            {45.458513, -73.640776},
-            {45.458487, -73.640709},
-            {45.458439, -73.640749},
-            {45.458464, -73.640814},
-            {45.458414, -73.640854},
-            {45.458473, -73.641007},
-            {45.458383, -73.641077},
-            {45.458428, -73.641199},
-            {45.458511, -73.641137},
-            {45.458548, -73.641232},
-            {45.458526, -73.641245},
-            {45.458537, -73.641278},
-            {45.458486, -73.641317}
-        };
-    
-        // (SC) 7141, Sherbrooke O. (Student Center)
-        double[][] lSCbuilding =
-        {
-            {45.459109, -73.639425},
-            {45.459226, -73.639336},
-            {45.459244, -73.639385},
-            {45.459310, -73.639330},
-            {45.459271, -73.639230},
-            {45.459293, -73.639210},
-            {45.459205, -73.638977},
-            {45.459157, -73.639001},
-            {45.459141, -73.638962},
-            {45.459076, -73.639010},
-            {45.459089, -73.639055},
-            {45.458997, -73.639130}
-        };
-    
-        // (SH) 7141, Sherbrooke O. (Solar House)
-        double[][] lSHbuilding =
-        {
-            {45.459257, -73.642559},
-            {45.459327, -73.642556},
-            {45.459323, -73.642372},
-            {45.459252, -73.642374}
-        };
-    
-        // (SI) 4455, West Broadway (St. Ignatius of Loyola Church)
-        double[][] lSIbuilding =
-        {
-            {45.457965, -73.642684},
-            {45.458167, -73.642525},
-            {45.458116, -73.642399},
-            {45.458097, -73.642414},
-            {45.458040, -73.642266},
-            {45.458052, -73.642234},
-            {45.457995, -73.642082},
-            {45.457969, -73.642082},
-            {45.457940, -73.642008},
-            {45.457878, -73.642053},
-            {45.457856, -73.641996},
-            {45.457823, -73.642024},
-            {45.457833, -73.642049},
-            {45.457765, -73.642104},
-            {45.457777, -73.642135},
-            {45.457721, -73.642180},
-            {45.457729, -73.642200},
-            {45.457679, -73.642241},
-            {45.457689, -73.642270},
-            {45.457648, -73.642302},
-            {45.457643, -73.642291},
-            {45.457633, -73.642299},
-            {45.457640, -73.642319},
-            {45.457622, -73.642333},
-            {45.457630, -73.642353},
-            {45.457569, -73.642399},
-            {45.457631, -73.642574},
-            {45.457655, -73.642557},
-            {45.457658, -73.642564},
-            {45.457729, -73.642507},
-            {45.457739, -73.642533},
-            {45.457781, -73.642501},
-            {45.457794, -73.642533},
-            {45.457844, -73.642492},
-            {45.457856, -73.642521},
-            {45.457911, -73.642476},
-            {45.457926, -73.642515},
-            {45.457931, -73.642511},
-            {45.457941, -73.642537},
-            {45.457916, -73.642558}
-        };
-    
-        // (SP) 7141, Sherbrooke O. (Richard J. Renaud Science Complex)
-        double[][] lSPbuilding =
-        {
-            {45.457438, -73.642003},
-            {45.457642, -73.641846},
-            {45.457672, -73.641928},
-            {45.458327, -73.641412},
-            {45.458276, -73.641283},
-            {45.458208, -73.641337},
-            {45.458180, -73.641262},
-            {45.458256, -73.641201},
-            {45.458194, -73.641037},
-            {45.458339, -73.640922},
-            {45.458316, -73.640862},
-            {45.457998, -73.641114},
-            {45.457979, -73.641066},
-            {45.457893, -73.641132},
-            {45.457908, -73.641170},
-            {45.457524, -73.641472},
-            {45.457202, -73.640657},
-            {45.456983, -73.640829},
-            {45.457024, -73.640935},
-            {45.456996, -73.640960},
-            {45.457017, -73.641015},
-            {45.457041, -73.640992},
-            {45.457158, -73.641295},
-            {45.457148, -73.641303},
-            {45.457179, -73.641382},
-            {45.457169, -73.641392},
-            {45.457185, -73.641432},
-            {45.457210, -73.641414}
-        };
-    
-        // (TA) 7079, de Terrebonne
-        double[][] lTAbuilding =
-        {
-            {45.459982, -73.640990},
-            {45.460031, -73.640950},
-            {45.459998, -73.640851},
-            {45.459943, -73.640890}
-        };
-    
-        // (TB) 7075, de Terrebonne
-        double[][] lTBbuilding =
-        {
-            {45.460031, -73.640950},
-            {45.460087, -73.640906},
-            {45.460049, -73.640809},
-            {45.459998, -73.640851}
-        };
-    
-        // (VE) 7141, Sherbrooke O. (Vanier extension)
-        double[][] lVEbuilding =
-        {
-            {45.458836, -73.639003},
-            {45.459086, -73.638803},
-            {45.458887, -73.638292},
-            {45.458629, -73.638469}
-        };
-    
-        // (VL) 7141, Sherbrooke O. (Vanier Library Building)
-        double[][] lVLbuilding =
-        {
-            {45.459168, -73.638807},
-            {45.459323, -73.638688},
-            {45.459146, -73.638226},
-            {45.459221, -73.638167},
-            {45.459107, -73.637872},
-            {45.458907, -73.638026},
-            {45.458914, -73.638051},
-            {45.458883, -73.638079},
-            {45.458899, -73.638121},
-            {45.458869, -73.638146},
-            {45.458889, -73.638195},
-            {45.458856, -73.638221},
-            {45.458886, -73.638297},
-            {45.459087, -73.638800},
-            {45.459151, -73.638767}
-        };
-        
-        return new double[][][]{
-        lADbuilding, lBBbuilding, lBHbuilding, lCCbuilding, lCJbuilding, lDObuilding,
-        lFCbuilding, lGEbuilding, lHAbuilding, lHBbuilding, lHCbuilding, lHUbuilding,
-        lJRbuilding, lPCbuilding, lPSbuilding, lPTbuilding, lPYbuilding, lRAbuilding,
-        lRFbuilding, lSCbuilding, lSHbuilding, lSIbuilding, lSPbuilding, lTAbuilding,
-        lTBbuilding, lVEbuilding, lVLbuilding};
-    }
-}
